# frontend/moderator-dashboard-ui/docs/verification/backend/openapi_verified.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B2B Platform API",
    "description": "API for B2B Platform - supplier moderation and parsing system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/moderator/tasks": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Moderator Tasks",
        "operationId": "list_moderator_tasks_moderator_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/moderator/dashboard-stats": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Moderator Dashboard Stats",
        "operationId": "get_moderator_dashboard_stats_moderator_dashboard_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeratorDashboardStatsDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/moderator/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Users",
        "operationId": "list_users_moderator_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserAccessDTO"
                  },
                  "type": "array",
                  "title": "Response List Users Moderator Users Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "Create user with email (for Yandex OAuth access).",
        "operationId": "create_user_moderator_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/moderator/users/{user_id}/cabinet-access": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update User Cabinet Access",
        "operationId": "update_user_cabinet_access_moderator_users__user_id__cabinet_access_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCabinetAccessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccessDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/moderator/users/{user_id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User",
        "operationId": "delete_user_moderator_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "CreateUserRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModeratorDashboardStatsDTO": {
        "properties": {
          "domains_in_queue": {
            "type": "integer",
            "title": "Domains In Queue"
          },
          "enrichment_domains_in_queue": {
            "type": "integer",
            "title": "Enrichment Domains In Queue"
          },
          "new_suppliers": {
            "type": "integer",
            "title": "New Suppliers"
          },
          "new_suppliers_week": {
            "type": "integer",
            "title": "New Suppliers Week"
          },
          "active_runs": {
            "type": "integer",
            "title": "Active Runs"
          },
          "blacklist_count": {
            "type": "integer",
            "title": "Blacklist Count"
          },
          "open_tasks": {
            "type": "integer",
            "title": "Open Tasks"
          }
        },
        "type": "object",
        "required": [
          "domains_in_queue",
          "enrichment_domains_in_queue",
          "new_suppliers",
          "new_suppliers_week",
          "active_runs",
          "blacklist_count",
          "open_tasks"
        ],
        "title": "ModeratorDashboardStatsDTO"
      },
...
      "UpdateCabinetAccessRequest": {
        "properties": {
          "cabinet_access_enabled": {
            "type": "boolean",
            "title": "Cabinet Access Enabled"
          }
        },
        "type": "object",
        "required": [
          "cabinet_access_enabled"
        ],
        "title": "UpdateCabinetAccessRequest"
      },
...
      "UserAccessDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "cabinet_access_enabled": {
            "type": "boolean",
            "title": "Cabinet Access Enabled"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "is_active",
          "cabinet_access_enabled"
        ],
        "title": "UserAccessDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}